Added dac-adc support
Now able to identify & store DAC-ADC type devices.

One thing to note is that DAC-ADC devices currently have non-unique
responses to the *IDN? command, meaning that having multiple of them
active at once would confuse the Serial Device Manager. For now that is
not a problem, but unique names in the IDN response would be a good
idea.